Will add containers and css modules for both coming soon and top rated pages. Will renamed files and components from Movies to Home
