+++ client/components/Navbar/Navbar.jsx
@@ -32,7 +32,7 @@ const Navbar = () => {
           { (focused && searchResult) && <SearchResult searchResult={searchResult} /> }
           { error && <p>{error}</p>}
       </div>
-      <Link className={css.navlink} to='/upcoming'>Coming Soon</Link>
+      <Link className={css.navlink} to='/coming-soon'>Coming Soon</Link>
       <Link className={css.navlink} to='/top-rated'>Top Rated</Link>
       <Link className={css.end} to='login'>Sign In</Link>
     </nav>
