+++ client/containers/ComingSoon.jsx
@@ -1,0 +1,12 @@
+import React, { useState } from 'react';
+import css from '../styles/ComingSoon.module.css';
+
+const ComingSoon = () => {
+
+  return ( 
+  <div className={css.comingSoon}>
+    <div className={css.innerCS}></div>
+  </div> );
+}
+ 
+export default ComingSoon;
