+++ client/containers/TopRated.jsx
@@ -1,0 +1,11 @@
+import React, { useState } from 'react';
+import css from '../styles/TopRated.module.css'
+const TopRated = () => {
+  return ( 
+    <div className={css.topRated}>
+
+    </div>
+   );
+}
+ 
+export default TopRated;
