+++ client/styles/Preview.module.css
@@ -138,6 +138,7 @@ h2 {
 
 .previewClose:hover {
   background-color: rgba(255, 255, 255, 0.219);
+  cursor: pointer;
 }
 
 .previewOverview {
