+++ client/styles/TopRated.module.css
@@ -1,0 +1,3 @@
+.topRated {
+  border: solid white 2px;
+}
